refactor(site): adding a few things left out of merge
