+++ configurations/default/settings.yml
@@ -9,5 +9,5 @@ geocoder:
 s3bucket: commute-staging
 modeshareStats:
   carpool: 0.10
-  other: 0.02
+  bike: 0.02
   transit: 0.14
