+++ package.json
@@ -94,7 +94,7 @@
     "react-formal-inputs": "^2.0.0",
     "react-leaflet": "^1.1.0",
     "react-leaflet-heatmap-layer": "https://github.com/evansiroky/react-leaflet-heatmap-layer.git",
-    "react-modal": "^2.2.1",
+    "react-modal": "^2.2.4",
     "react-redux": "^4.4.5",
     "react-router": "^3.0.0",
     "react-router-redux": "^4.0.6",
